+++ index.html
@@ -1,3 +1,4 @@
+<script src="js/main.js" type="module"></script>
 <!DOCTYPE html>
 <html lang="ru">
 <head>
@@ -6,7 +7,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="main.js"></script>
   <title>Кекстаграм</title>
 </head>
